add workspace permission

+++ Ticket.API/Controllers/WorkSpacesController.cs
@@ -5,11 +5,13 @@
     [Route("api/[controller]")]
     public class WorkSpacesController : BaseController
     {
+        private readonly IAuthorizationService _authService;
         private readonly IWorkSpaceService _workSpaceService;
         private readonly IMapper _mapper;
 
-        public WorkSpacesController(IWorkSpaceService workSpaceService, IMapper mapper)
+        public WorkSpacesController(IAuthorizationService authService, IWorkSpaceService workSpaceService, IMapper mapper)
         {
+            _authService = authService;
             _workSpaceService = workSpaceService;
             _mapper = mapper;
         }
@@ -22,6 +24,10 @@
         [HttpGet]
         public async Task<BaseResponseWithPagination<List<WorkSpaceResponseModel>>> GetAllWorkSpace([FromQuery] WorkSpaceRequestModel model)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.GetListWorkSpace);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             var res = await _workSpaceService.GetWorkSpaces(model);
             return SuccessWithPagination(res.Pagination, res.WorkSpaces);
         }
@@ -34,6 +40,10 @@
         [HttpPost]
         public async Task<BaseResponse> CreateNewWorkSpace([FromBody] WorkSpaceCreateRequestModel model)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.CreateWorkSpace);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             await _workSpaceService.CreateWorkSpace(_mapper.Map<WorkSpaceCreateMapRequestModel>(model), User.Identity.Name);
             return Success();
         }
@@ -47,6 +57,10 @@
         [HttpPut("{workSpaceId}")]
         public async Task<BaseResponse> UpdateExistWorkSpace([FromRoute] string workSpaceId, [FromBody] WorkSpaceUpdateRequestModel model)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.UpdateWorkSpace);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             await _workSpaceService.UpdateWorkSpace(_mapper.Map<WorkSpaceUpdateMapRequestModel>(model), User.Identity.Name, workSpaceId);
             return Success();
         }
@@ -59,6 +73,10 @@
         [HttpDelete("{workSpaceId}")]
         public async Task<BaseResponse> DeleteExistWorkSpace([FromRoute] string workSpaceId)
         {
+            var result = await _authService.AuthorizeAsync(User, ApplicationPermissions.DeleteWorkSpace);
+            if (!result.Succeeded)
+                throw new BaseException(ErrorCodes.FORBIDDEN, HttpCodes.FOR_BIDDEN, ErrorCodes.FORBIDDEN.GetEnumMemberValue());
+
             await _workSpaceService.DeleteWorkSpace(workSpaceId, User.Identity.Name);
             return Success();
         }
